Make temporary fix for Modello navigator.
Take into account that this fix will make all
synthetic speech coming from TTS behaving like
it would come from navigator. That means that
it is attenuating the underlying music stream and
mixing with it. This might actually work for most
TTS cases, but not for all. Eventually this has
to be fixed in Winthorpe and xwalk TTS plugin.

Change-Id: Ie0ee1c7efa99a757a0cbb8002e1fae627ff535ca