add input routing group "default_driver"
This adds an input counterpart to the output routing group
"default_driver". I want to be able to test things with a simple
parecord command, and this is required (but not sufficient) for making
that work properly. Initial routing (a.k.a. prerouting) already
worked, but rerouting when new sources appear didn't work, because
(among other problems) there was no appropriate routing group defined
parecord.

Change-Id: I20795d73d853627cd1d9746b5fa4a7412f83970c