remove default case from mir_node_type_str()
Having a default case prevents the compiler from warning about
unhandled cases, which can easily cause cases to be missed when new
members are added to the enum. In this cases there were four node
types that weren't handled in the switch (causing them to be reported
as user defined types): null, front speakers, rear speakers and usb
headphone.

Change-Id: Ia1c24b4ffc8dff52e7c259a5a4f5f01ece34bad3