add the "unspecified_output_stream" application class
This is similar in use to the "player" class for input streams, i.e.
used as the fallback when no other class matches a stream. For
example, parecord without any media role will be assigned to this
class.

I'm adding this, because without an application class, parecord
doesn't get rerouted when new sources appear (in my case I want to
route to a tunnel source when it appears).

Change-Id: Ied6c83a3a9e23b405320d5fe538d232624d03499